feat: edit gh-ost flags
