+++ frontend/src/components/IssueV1/components/StageSection/Actions/GhostSection/common.ts
@@ -1,4 +1,5 @@
 import { InjectionKey, Ref, computed, inject, provide, ref } from "vue";
+import { useI18n } from "vue-i18n";
 import {
   databaseForTask,
   getLocalSheetByName,
@@ -7,17 +8,24 @@ import {
   specForTask,
   useIssueContext,
 } from "@/components/IssueV1/logic";
+import { useCurrentUserV1 } from "@/store";
 import { ComposedDatabase, ComposedIssue } from "@/types";
 import { Engine } from "@/types/proto/v1/common";
+import { IssueStatus } from "@/types/proto/v1/issue_service";
 import {
   Plan_ChangeDatabaseConfig_Type,
   Plan_Spec,
   Task,
   Task_Status,
+  Task_Type,
+  task_StatusToJSON,
 } from "@/types/proto/v1/rollout_service";
 import {
   MIN_GHOST_SUPPORT_MYSQL_VERSION,
+  extractUserResourceName,
+  flattenTaskV1List,
   getSheetStatement,
+  hasWorkspacePermissionV1,
   semverCompare,
 } from "@/utils";
 
@@ -26,6 +34,7 @@ export type GhostUIViewType = "NONE" | "OFF" | "ON";
 export type IssueGhostContext = {
   viewType: Ref<GhostUIViewType>;
   showFlagsPanel: Ref<boolean>;
+  denyEditGhostFlagsReasons: Ref<string[]>;
   toggleGhost: (spec: Plan_Spec, on: boolean) => Promise<void>;
 };
 
@@ -38,12 +47,79 @@ export const useIssueGhostContext = () => {
 };
 
 export const provideIssueGhostContext = () => {
-  const { issue, selectedTask: task, reInitialize } = useIssueContext();
+  const { t } = useI18n();
+  const me = useCurrentUserV1();
+  const {
+    isCreating,
+    issue,
+    selectedTask: task,
+    reInitialize,
+  } = useIssueContext();
 
   const viewType = computed((): GhostUIViewType => {
     return ghostViewTypeForTask(issue.value, task.value);
   });
   const showFlagsPanel = ref(false);
+  const isDeploymentConfig = computed(() => {
+    const spec = specForTask(issue.value.planEntity, task.value);
+    return !!spec?.changeDatabaseConfig?.target?.match(
+      /\/deploymentConfigs\/[^/]+/
+    );
+  });
+
+  const denyEditGhostFlagsReasons = computed(() => {
+    if (isCreating.value) {
+      return [];
+    }
+
+    if (issue.value.status !== IssueStatus.OPEN) {
+      return [t("issue.error.issue-is-not-open")];
+    }
+
+    const errors: string[] = [];
+
+    if (extractUserResourceName(issue.value.creator) !== me.value.email) {
+      if (
+        !hasWorkspacePermissionV1(
+          "bb.permission.workspace.manage-issue",
+          me.value.userRole
+        )
+      ) {
+        return [t("issue.error.you-don-have-privilege-to-edit-this-issue")];
+      }
+    }
+
+    if (isDeploymentConfig.value) {
+      // If a task is created by deploymentConfig. It is editable only when all
+      // its "brothers" (created by the same deploymentConfig) are editable.
+      // By which "editable" means a task's status meets the requirements.
+      const ghostSyncTasks = flattenTaskV1List(
+        issue.value.rolloutEntity
+      ).filter(
+        (task) => task.type === Task_Type.DATABASE_SCHEMA_UPDATE_GHOST_SYNC
+      );
+      if (
+        ghostSyncTasks.some(
+          (task) => !allowChangeTaskGhostFlags(issue.value, task)
+        )
+      ) {
+        errors.push(
+          t(
+            "task.online-migration.error.some-tasks-are-not-editable-in-batch-mode"
+          )
+        );
+      }
+    } else {
+      if (!allowChangeTaskGhostFlags(issue.value, task.value)) {
+        errors.push(
+          t("task.online-migration.error.x-status-task-is-not-editable", {
+            status: task_StatusToJSON(task.value.status),
+          })
+        );
+      }
+    }
+    return errors;
+  });
 
   const toggleGhost = async (spec: Plan_Spec, on: boolean) => {
     const overrides: Record<string, string> = {};
@@ -69,6 +145,7 @@ export const provideIssueGhostContext = () => {
   const context: IssueGhostContext = {
     viewType,
     showFlagsPanel,
+    denyEditGhostFlagsReasons,
     toggleGhost,
   };
